Multi Class Texture Analysis In Colorectal Cancer Histology
Introduction
Multi-class texture analysis in colorectal cancer histology

Automatic recognition of different tissue types in histological images is an essential part in the digital pathology toolbox. Texture analysis is commonly used to address this problem; mainly in the context of estimating the tumour/stroma ratio on histological samples.

In this paper we present a new dataset of 5,000 histological images of human colorectal cancer including eight different types of tissue.

We used this set to assess the classification performance of a wide range of texture descriptors and classification.